fix: OBJECT_MAX_SIZEとファイル行数が同じ場合や末尾に改行がある場合にエラーになってしまう問題を解決。
#50 #59
