+++ includes/data.h
@@ -21,7 +21,7 @@ kawadaさんがEPSILONは 0.000001 くらいがいいって言ってた
 # define EPSILON 0.000001
 
 # ifndef OBJECT_SIZE_MAX
-#  define OBJECT_SIZE_MAX 4200
+#  define OBJECT_SIZE_MAX 42000
 # endif
 
 typedef struct s_vect
