+++ includes/loader.h
@@ -7,6 +7,7 @@
 # include <string.h>
 # include <stdlib.h>
 # include <errno.h>
+# include <sys/stat.h>
 
 void	load_file(t_data *data, char *filepath);
 void	assign_line_to_object(char *line, t_object *obj);
