+++ sample/box.tri
